move test, refactor dir
