+++ zkevm-circuits/src/evm_circuit/execution/callop.rs
@@ -1489,4 +1489,328 @@ mod test {
             })
             .run();
     }
+
+    #[test]
+    fn test_precompiled_call() {
+        use bus_mapping::{
+            mock::BlockData,
+            precompile::PrecompileCallArgs,
+        };
+        use eth_types::{bytecode, evm_types::OpcodeId, geth_types::GethData, word, Word};
+        use mock::{
+            test_ctx::{
+                helpers::{account_0_code_account_1_no_code, tx_from_1_to_0},
+                LoggerConfig,
+            },
+            TestContext,
+        };
+
+        let test_vector = [
+            PrecompileCallArgs {
+                name: "ecRecover",
+                setup_code: bytecode! {
+                    PUSH32(word!("0x456e9aea5e197a1f1af7a3e85a3212fa4049a3ba34c2289b4c860fc0b0c64ef3")) // hash
+                    PUSH1(0x0)
+                    MSTORE
+                    PUSH1(28) // v
+                    PUSH1(0x20)
+                    MSTORE
+                    PUSH32(word!("0x9242685bf161793cc25603c231bc2f568eb630ea16aa137d2664ac8038825608")) // r
+                    PUSH1(0x40)
+                    MSTORE
+                    PUSH32(word!("0x4f8ae3bd7535248d0bd448298cc2e2071e56992d0774dc340c368ae950852ada")) // s
+                    PUSH1(0x60)
+                    MSTORE
+                },
+                ret_size: Word::from(0x20),
+                ret_offset: Word::from(0x80),
+                call_data_length: Word::from(0x80),
+                address: Word::from(0x1),
+                stack_value: vec![(
+                    Word::from(0x80),
+                    word!("7156526fbd7a3c72969b54f64e42c10fbb768c8a"),
+                )],
+                ..Default::default()
+            },
+            PrecompileCallArgs {
+                name: "SHA2-256",
+                setup_code: bytecode! {
+                    PUSH1(0xFF) // data
+                    PUSH1(0)
+                    MSTORE
+                },
+                ret_size: Word::from(0x20),
+                ret_offset: Word::from(0x20),
+                call_data_length: Word::from(0x1),
+                call_data_offset: Word::from(0x1F),
+                address: Word::from(0x2),
+                stack_value: vec![(
+                    Word::from(0x20),
+                    word!("a8100ae6aa1940d0b663bb31cd466142ebbdbd5187131b92d93818987832eb89"),
+                )],
+                ..Default::default()
+            },
+            PrecompileCallArgs {
+                name: "RIPEMD-160",
+                setup_code: bytecode! {
+                    PUSH1(0xFF) // data
+                    PUSH1(0)
+                    MSTORE
+                },
+                ret_size: Word::from(0x20),
+                ret_offset: Word::from(0x20),
+                call_data_length: Word::from(0x1),
+                call_data_offset: Word::from(0x1F),
+                address: Word::from(0x3),
+                stack_value: vec![(
+                    Word::from(0x20),
+                    word!("2c0c45d3ecab80fe060e5f1d7057cd2f8de5e557"),
+                )],
+                ..Default::default()
+            },
+            PrecompileCallArgs {
+                name: "identity",
+                setup_code: bytecode! {
+                    PUSH16(word!("0123456789ABCDEF0123456789ABCDEF"))
+                    PUSH1(0x00)
+                    MSTORE
+                },
+                ret_size: Word::from(0x20),
+                ret_offset: Word::from(0x20),
+                call_data_length: Word::from(0x20),
+                address: Word::from(0x4),
+                stack_value: vec![(Word::from(0x20), word!("0123456789ABCDEF0123456789ABCDEF"))],
+                ..Default::default()
+            },
+            // PrecompileCallArgs {
+            // name: "modexp",
+            // setup_code: bytecode! {
+            // PUSH1(1) // Bsize
+            // PUSH1(0)
+            // MSTORE
+            // PUSH1(1) // Esize
+            // PUSH1(0x20)
+            // MSTORE
+            // PUSH1(1) // Msize
+            // PUSH1(0x40)
+            // MSTORE
+            // PUSH32(word!("0x08090A0000000000000000000000000000000000000000000000000000000000"))
+            // // B, E and M PUSH1(0x60)
+            // MSTORE
+            // },
+            // ret_size: Word::from(0x01),
+            // ret_offset: Word::from(0x9F),
+            // call_data_length: Word::from(0x63),
+            // address: Word::from(0x5),
+            // stack_value: vec![(Word::from(0x80), Word::from(8))],
+            // ..Default::default()
+            // },
+            PrecompileCallArgs {
+                name: "ecAdd",
+                setup_code: bytecode! {
+                    PUSH1(1) // x1
+                    PUSH1(0)
+                    MSTORE
+                    PUSH1(2) // y1
+                    PUSH1(0x20)
+                    MSTORE
+                    PUSH1(1) // x2
+                    PUSH1(0x40)
+                    MSTORE
+                    PUSH1(2) // y2
+                    PUSH1(0x60)
+                    MSTORE
+                },
+                ret_size: Word::from(0x40),
+                ret_offset: Word::from(0x80),
+                call_data_length: Word::from(0x80),
+                address: Word::from(0x6),
+                stack_value: vec![
+                    (
+                        Word::from(0x80),
+                        word!("30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd3"),
+                    ),
+                    (
+                        Word::from(0xA0),
+                        word!("15ed738c0e0a7c92e7845f96b2ae9c0a68a6a449e3538fc7ff3ebf7a5a18a2c4"),
+                    ),
+                ],
+                ..Default::default()
+            },
+            PrecompileCallArgs {
+                name: "ecMul",
+                setup_code: bytecode! {
+                    PUSH1(1) // x1
+                    PUSH1(0)
+                    MSTORE
+                    PUSH1(2) // y1
+                    PUSH1(0x20)
+                    MSTORE
+                    PUSH1(2) // s
+                    PUSH1(0x40)
+                    MSTORE
+                },
+                ret_size: Word::from(0x40),
+                ret_offset: Word::from(0x60),
+                call_data_length: Word::from(0x60),
+                address: Word::from(0x7),
+                stack_value: vec![
+                    (
+                        Word::from(0x60),
+                        word!("30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd3"),
+                    ),
+                    (
+                        Word::from(0x80),
+                        word!("15ed738c0e0a7c92e7845f96b2ae9c0a68a6a449e3538fc7ff3ebf7a5a18a2c4"),
+                    ),
+                ],
+                ..Default::default()
+            },
+            // PrecompileCallArgs {
+            // name: "ecPairing",
+            // setup_code: bytecode! {
+            // PUSH32(word!("0x23a8eb0b0996252cb548a4487da97b02422ebc0e834613f954de6c7e0afdc1fc"))
+            // PUSH32(word!("0x2a23af9a5ce2ba2796c1f4e453a370eb0af8c212d9dc9acd8fc02c2e907baea2"))
+            // PUSH32(word!("0x091058a3141822985733cbdddfed0fd8d6c104e9e9eff40bf5abfef9ab163bc7"))
+            // PUSH32(word!("0x1971ff0471b09fa93caaf13cbf443c1aede09cc4328f5a62aad45f40ec133eb4"))
+            // PUSH32(word!("0x30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd45"))
+            // PUSH32(word!("0x0000000000000000000000000000000000000000000000000000000000000001"))
+            // PUSH32(word!("0x2fe02e47887507adf0ff1743cbac6ba291e66f59be6bd763950bb16041a0a85e"))
+            // PUSH32(word!("0x2bd368e28381e8eccb5fa81fc26cf3f048eea9abfdd85d7ed3ab3698d63e4f90"))
+            // PUSH32(word!("0x22606845ff186793914e03e21df544c34ffe2f2f3504de8a79d9159eca2d98d9"))
+            // PUSH32(word!("0x1fb19bb476f6b9e44e2a32234da8212f61cd63919354bc06aef31e3cfaff3ebc"))
+            // PUSH32(word!("0x2c0f001f52110ccfe69108924926e45f0b0c868df0e7bde1fe16d3242dc715f6"))
+            // PUSH32(word!("0x2cf44499d5d27bb186308b7af7af02ac5bc9eeb6a3d147c186b21fb1b76e18da"))
+            //
+            // PUSH1(12)
+            // PUSH2(0x200)
+            // MSTORE
+            //
+            // JUMPDEST
+            //
+            // PUSH2(0x200)
+            // MLOAD
+            // PUSH1(12)
+            // SUB
+            // PUSH1(0x20)
+            // MUL
+            // MSTORE
+            // PUSH1(1)
+            // PUSH2(0x200)
+            // MLOAD
+            // SUB
+            // DUP1
+            // PUSH2(0x200)
+            // MSTORE
+            // PUSH2(0x192)
+            // JUMPI
+            // },
+            // ret_size: Word::from(0x20),
+            // call_data_length: Word::from(0x180),
+            // address: Word::from(0x8),
+            // stack_value: vec![(Word::from(0x0), Word::from(1))],
+            // max_rws: 3000,
+            // ..Default::default()
+            // },
+            // PrecompileCallArgs {
+            // name: "blake2f",
+            // setup_code: bytecode! {
+            // PUSH32(word!("0000000003000000000000000000000000000000010000000000000000000000"))
+            // PUSH32(word!("0000000000000000000000000000000000000000000000000000000000000000"))
+            // PUSH32(word!("0000000000000000000000000000000000000000000000000000000000000000"))
+            // PUSH32(word!("0000000000000000000000000000000000000000000000000000000000000000"))
+            // PUSH32(word!("19cde05b61626300000000000000000000000000000000000000000000000000"))
+            // PUSH32(word!("3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e13"))
+            // PUSH32(word!("0000000048c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f"))
+            //
+            // PUSH1(7)
+            // PUSH2(0x160)
+            // MSTORE
+            //
+            // JUMPDEST
+            //
+            // PUSH2(0x160)
+            // MLOAD
+            // PUSH1(7)
+            // SUB
+            // PUSH1(0x20)
+            // MUL
+            // MSTORE
+            // PUSH1(1)
+            // PUSH2(0x160)
+            // MLOAD
+            // SUB
+            // DUP1
+            // PUSH2(0x160)
+            // MSTORE
+            // PUSH2(0xed)
+            // JUMPI
+            // },
+            // ret_size: Word::from(0x40),
+            // ret_offset: Word::from(0x0),
+            // call_data_length: Word::from(0xd5),
+            // address: Word::from(0x9),
+            // stack_value: vec![
+            // (
+            // Word::from(0x20),
+            // #[cfg(feature = "scroll")]
+            // word!("3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e13"),
+            // #[cfg(not(feature = "scroll"))]
+            // word!("d282e6ad7f520e511f6c3e2b8c68059b9442be0454267ce079217e1319cde05b"),
+            // ),
+            // (
+            // Word::from(0x0),
+            // #[cfg(feature = "scroll")]
+            // word!("0000000048c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f"),
+            // #[cfg(not(feature = "scroll"))]
+            // word!("8c9bcf367e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5"),
+            // ),
+            // ],
+            // max_rws: 3000,
+            // ..Default::default()
+            // },
+        ];
+
+        let call_ops = [
+            OpcodeId::CALL,
+            OpcodeId::CALLCODE,
+            OpcodeId::DELEGATECALL,
+            OpcodeId::STATICCALL,
+        ];
+
+        for (test_call, call_op) in itertools::iproduct!(test_vector.iter(), call_ops.iter()) {
+            let code = test_call.with_call_op(*call_op);
+            let block: GethData = TestContext::<2, 1>::new_with_logger_config(
+                None,
+                account_0_code_account_1_no_code(code),
+                tx_from_1_to_0,
+                |block, _tx| block.number(0xcafeu64),
+                LoggerConfig {
+                    enable_memory: true,
+                    ..Default::default()
+                },
+            )
+            .unwrap()
+            .into();
+
+            let builder = BlockData::new_from_geth_data(block.clone()).new_circuit_input_builder();
+            builder
+                .handle_block(&block.eth_block, &block.geth_traces)
+                .unwrap();
+
+            let step = block.geth_traces[0]
+                .struct_logs
+                .last()
+                .expect("at least one step");
+            log::debug!("{:?}", step.stack);
+            for (offset, (_, stack_value)) in test_call.stack_value.iter().enumerate() {
+                assert_eq!(
+                    *stack_value,
+                    step.stack.nth_last(offset).expect("stack value not found"),
+                    "stack output mismatch {}",
+                    test_call.name
+                );
+            }
+        }
+    }
 }
